test(cmd): consolidate tidy tests into tidy_test.go
